Moved the DatePicker component to the common components.
