+++ src/javascript/query-explorer/components/query-explorer.js
@@ -16,13 +16,13 @@
 import React from 'react';
 import accountSummaries from 'javascript-api-utils/lib/account-summaries';
 
-import Datepicker from './datepicker';
 import QueryReport from './query-report';
 import Select2MultiSuggest from './select2-multi-suggest';
 import ViewSelector from './view-selector';
 
 import {gaAll} from '../../analytics';
 import AlertDispatcher from '../../components/alert-dispatcher';
+import Datepicker from '../../components/datepicker';
 import HelpIconLink from '../../components/help-icon-link';
 import SearchSuggest from '../../components/search-suggest';
